Fix data type on post meta value

+++ src/Tracker/HitRecorder.php
@@ -15,11 +15,11 @@ class HitRecorder {
 		$hit_key = 'hit_' . md5( $post_id . Helpers::client_ip() . date( 'Y-m-d' ) );
 		set_transient( $hit_key, 1, DAY_IN_SECONDS );
 
-		$total = get_post_meta( $post_id, 'post_hits', true );
+		$total = (int) get_post_meta( $post_id, 'post_hits', true );
 		update_post_meta( $post_id, 'post_hits', $total + 1 );
 
 		$daily_key = 'post_hits_daily_' . date( 'Y-m-d' );
-		$daily     = get_post_meta( $post_id, $daily_key, true );
+		$daily     = (int) get_post_meta( $post_id, $daily_key, true );
 		update_post_meta( $post_id, $daily_key, $daily + 1 );
 	}
 }
